+++ pkg/bievents/types.go
@@ -9,6 +9,7 @@ const (
 	Login             ActivityType = "login"
 	EIDLogin          ActivityType = "eid-login"
 	Logout            ActivityType = "logout"
+	TokenRefreshed    ActivityType = "token-refreshed"
 	PhoneVerify       ActivityType = "phone-verify"
 	DeviceRegister    ActivityType = "device-register"
 	DeviceDelete      ActivityType = "device-delete"
@@ -16,6 +17,7 @@ const (
 	LoginComplete     ActivityType = "login-complete"
 	SharingStart      ActivityType = "sharing-start"
 	SharingComplete   ActivityType = "sharing-complete"
+	SharingRevoked    ActivityType = "sharing-revoked"
 	DocumentDelete    ActivityType = "document-delete"
 	DocumentDeleteAll ActivityType = "document-delete-all"
 	DocumentUpload    ActivityType = "document-upload"
@@ -24,30 +26,35 @@ const (
 	RecordCreate      ActivityType = "record-create"
 )
 
-// OnboardingData is used to define details about onboarding data.
-// This type can be used to define structs
-type OnboardingData struct {
-	CUC         string    `json:"cuc"`
+type UserRegisterData struct {
+	CucID       string    `json:"cuc-id"`
 	AccountType EmailType `json:"account-type"`
 	SourceURL   string    `json:"source-url"`
 	Source      string    `json:"source"`
 }
 
 type LoginData struct {
-	ClientID string `json:"client-id"`
+	ClientID  string `json:"client-id"`
+	SourceURL string `json:"source-url"`
 }
 
 type EIDLoginData struct {
 	Challenge string `json:"eid-challenge"`
 	ClientID  string `json:"client-id"`
+	SourceURL string `json:"source-url"`
 }
 
 type LoginCompleteData struct {
-	SessionID string // a session identifier: it allows to connect a login and a logout event
+	SessionID string `json:"session-id"` // a session identifier: it allows to connect a login and a logout event
+	ClientID  string `json:"client-id"`
 }
 
 type LogoutData struct {
-	SessionID string // a session identifier: it allows to connect a login and a logout event
+	SessionID string `json:"session-id"` // a session identifier: it allows to connect a login and a logout event
+}
+
+type TokenRefreshedData struct {
+	SessionID string `json:"session-id"`
 }
 
 type EIDSamlEntranceData struct {
@@ -60,21 +67,25 @@ type DeviceRegisterData struct {
 	DeviceType string `json:"device-type"`
 	Challenge  string `json:"eid-challenge,omitempty"` // only for eID devices
 	DeviceID   string `json:"device-id"`
-	OwnerID    string `json:"owner-id"`
 }
 
 type DeviceDeleteData struct {
-	DeviceID string
+	DeviceID string `json:"device-id"`
 }
 
 type PhoneVerifyData struct {
-	DeviceID string
+	DeviceID string `json:"device-id"`
 }
 
 type SharingStartData struct {
-	SharingSessionID string
+	SharingSessionID string `json:"sharing-session-id"`
+	ClientID         string `json:"client-id"`
 }
 
 type SharingCompleteData struct {
-	SharingSessionID string
+	SharingSessionID string `json:"sharing-session-id"`
+}
+
+type SharingRevokedData struct {
+	SharingSessionID string `json:"sharing-session-id"`
 }